+++ src/components/Inputs/Dropdown/DropdownOptions/DropdownOptions.tsx
@@ -2,10 +2,15 @@ import classNames from "classnames";
 import React, { ReactElement, useEffect, useRef, useState } from "react";
 
 import { getPosition } from "../../../../helpers";
-import { Position } from "../../../../helpers/getPosition/getPosition-types";
+import { CalculatedPosition, Position } from "../../../../helpers/getPosition/getPosition-types";
 import { DropdownOptionsProps } from "../types";
 
 import "./DropdownOptions.css";
+
+type DropdownOptionsStyles = {
+  opacity: number;
+  maxHeight: number;
+} & Position;
 
 export const DropdownOptions = <T,>({
   filterElement,
@@ -18,21 +23,46 @@ export const DropdownOptions = <T,>({
   classNamesObj,
 }: DropdownOptionsProps<T>) => {
   const ref = useRef<HTMLUListElement>(null);
-  const [position, setPosition] = useState<Position | { opacity: number } | undefined>({ opacity: 0 });
+  const [dropdownOptionsStyles, setDropdownOptionsStyles] = useState<DropdownOptionsStyles | undefined>({
+    top: 0,
+    left: 0,
+    width: 0,
+    calculatedPosition: CalculatedPosition.AUTO_BOTTOM,
+    opacity: 0,
+    maxHeight: 0,
+  });
 
   useEffect(() => {
-    if (!ref.current) {
-      return;
-    }
+    const element = ref.current;
 
-    setPosition(getPosition(filterElement, ref.current, { consumerHasParentWidth: true }));
+    const resizeObserver = new ResizeObserver(() => {
+      if (!element) {
+        return;
+      }
+
+      const position: Position = getPosition(filterElement, ref.current, { consumerHasParentWidth: true });
+
+      const children = Array.from(element.children) as HTMLLIElement[];
+
+      const totalHeight = children.slice(0, 6).reduce((sum, child) => sum + child.offsetHeight, 0);
+
+      const maxHeight = totalHeight;
+
+      setDropdownOptionsStyles({ ...position, maxHeight, opacity: 1 });
+    });
+
+    resizeObserver.observe(document.body);
+
+    return () => {
+      resizeObserver.disconnect();
+    };
   }, []);
 
   return (
     <ul
       ref={ref}
-      className={classNames("m-dropdown-options", classNamesObj?.dropdownOptions)}
-      style={position}
+      className={classNames("m-dropdown-options", "m-scroll slim-scroll", classNamesObj?.dropdownOptions)}
+      style={dropdownOptionsStyles}
       data-id={uniqueDropdownId}
     >
       {dropdownOptions && dropdownOptions.length > 0 ? (
@@ -42,7 +72,7 @@ export const DropdownOptions = <T,>({
               key={option[valueKey] as string}
               data-id={uniqueDropdownId}
               onClick={(e) => handleDropdownChange(e, option)}
-              className={classNames("m-dropdown-option", classNamesObj?.dropdownOption, {
+              className={classNames("m-dropdown-option", "truncate-text", classNamesObj?.dropdownOption, {
                 selected: option[valueKey] === value?.[valueKey],
               })}
             >