+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [[#45]](https://github.com/Marcin-Migdal/m-component-library/issues/45) **[PATCH]** Add Breadcrumb component
 - [[#65]](https://github.com/Marcin-Migdal/m-component-library/issues/65) **[PATCH]** Add IconField component
 - [[#99]](https://github.com/Marcin-Migdal/m-component-library/issues/99) **[PATCH]** Block scrolling when overlay is visible
+- [[#100]](https://github.com/Marcin-Migdal/m-component-library/issues/100) **[PATCH]** Limit dropdown menu max height
 
 ### Release date 25.10.2024 - Release 1.0.6